# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: 團購/代購訂單管理 API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /stores/{storeId}/orders/{orderId}/items/{itemId}/capture: {"post":{"operationId":"captureCatalogOrderItem","tags":["orders"],"parameters":[{"name":"storeId","in":"path","required":true,"schema":{"type":"integer","minimum":1,"maximum":2147483647,"multipleOf":1}},{"name":"orderId","in":"path","required":true,"schema":{"type":"integer","minimum":1,"maximum":2147483647,"multipleOf":1}},{"name":"itemId","in":"path","required":true,"schema":{"type":"integer","minimum":1,"maximum":2147483647,"multipleOf":1}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PricingV2PreviewInput"}}}},"responses":{"200":{"description":"Capture one PENDING item using server pricing","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Order"}}}},"400":{"description":"Invalid input"},"401":{"description":"Authentication required"},"403":{"description":"Owner required"},"404":{"description":"Same-store reference missing"},"409":{"description":"Immutable history"},"422":{"description":"Cost incomplete or amount outside bounds"}}}}
  /stores/{storeId}/catalog-sales: {"get":{"operationId":"getCatalogSales","tags":["catalog"],"parameters":[{"name":"storeId","in":"path","required":true,"schema":{"type":"integer","minimum":1,"maximum":2147483647,"multipleOf":1}},{"name":"ids","in":"query","required":true,"schema":{"type":"string","pattern":"^[1-9][0-9]*(,[1-9][0-9]*)*$"}}],"responses":{"200":{"description":"Batched formal general/VIP sales","content":{"application/json":{"schema":{"type":"object","additionalProperties":true}}}},"400":{"description":"Invalid input"},"401":{"description":"Authentication required"},"403":{"description":"Owner required"},"404":{"description":"Same-store reference missing"},"409":{"description":"Immutable history"},"422":{"description":"Cost incomplete or amount outside bounds"}}}}
  /stores/{storeId}/catalog-products/{catalogProductId}/sales-reference: {"get":{"operationId":"getCatalogSalesReference","tags":["catalog"],"parameters":[{"name":"storeId","in":"path","required":true,"schema":{"type":"integer","minimum":1,"maximum":2147483647,"multipleOf":1}},{"name":"catalogProductId","in":"path","required":true,"schema":{"type":"integer","minimum":1,"maximum":2147483647,"multipleOf":1}}],"responses":{"200":{"description":"Historical stats with separate current-cost reference","content":{"application/json":{"schema":{"type":"object","additionalProperties":true}}}},"400":{"description":"Invalid input"},"401":{"description":"Authentication required"},"403":{"description":"Owner required"},"404":{"description":"Same-store reference missing"},"409":{"description":"Immutable history"},"422":{"description":"Cost incomplete or amount outside bounds"}}}}
  /stores/{storeId}/order-cost-options: {"get":{"operationId":"getCatalogOrderCostOptions","tags":["orders"],"parameters":[{"name":"storeId","in":"path","required":true,"schema":{"type":"integer","minimum":1,"maximum":2147483647,"multipleOf":1}}],"responses":{"200":{"description":"Explicit same-store trips and routes","content":{"application/json":{"schema":{"type":"object","additionalProperties":true}}}},"400":{"description":"Invalid input"},"401":{"description":"Authentication required"},"403":{"description":"Owner required"},"404":{"description":"Same-store reference missing"},"409":{"description":"Immutable history"},"422":{"description":"Cost incomplete or amount outside bounds"}}}}
  /stores/{storeId}/catalog-orders: {"post":{"operationId":"createCatalogOrder","tags":["orders"],"parameters":[{"name":"storeId","in":"path","required":true,"schema":{"type":"integer","minimum":1,"maximum":2147483647,"multipleOf":1}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CatalogOrderInput"}}}},"responses":{"201":{"description":"Created order","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Order"}}}}}}}
...
  /stores/{storeId}/products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products for a store
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /stores/{storeId}/products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
...
  schemas:
...
    CatalogOrderLineInput: {"type":"object","additionalProperties":false,"required":["quantity"],"properties":{"listingProductId":{"type":"integer","minimum":1,"maximum":2147483647,"multipleOf":1},"catalogProductId":{"type":"integer","minimum":1,"maximum":2147483647,"multipleOf":1},"name":{"type":"string","maxLength":200,"minLength":1},"quantity":{"type":"integer","minimum":1,"maximum":2147483647,"multipleOf":1},"unitPriceTwd":{"type":"string","pattern":"^\\d{1,8}(?:\\.\\d{1,2})?$"},"specValues":{"type":"object","additionalProperties":{"type":"string","maxLength":500}},"cost":{"$ref":"#/components/schemas/PricingV2PreviewInput"},"saveCatalog":{"$ref":"#/components/schemas/CatalogCreateInput"}}}
    CatalogOrderInput: {"type":"object","additionalProperties":false,"required":["items","buyerName","buyerPhone","pickupMethod"],"properties":{"items":{"type":"array","minItems":1,"maxItems":100,"items":{"$ref":"#/components/schemas/CatalogOrderLineInput"}},"buyerName":{"type":"string","maxLength":200,"minLength":1},"buyerPhone":{"type":"string","maxLength":100,"minLength":1},"pickupMethod":{"type":"string","maxLength":100,"minLength":1},"customerId":{"type":["integer","null"],"minimum":1,"maximum":2147483647,"multipleOf":1},"notes":{"type":["string","null"],"maxLength":5000},"creditSpent":{"type":"string","maxLength":100},"paymentLast5":{"type":["string","null"],"maxLength":5},"shippingMethod":{"type":["string","null"],"enum":["self_pickup","convenience_store","home_delivery","other",null]},"recipientName":{"type":["string","null"],"maxLength":1000},"recipientPhone":{"type":["string","null"],"maxLength":1000},"recipientAddress":{"type":["string","null"],"maxLength":1000},"storeCode":{"type":["string","null"],"maxLength":1000},"storeName":{"type":["string","null"],"maxLength":1000},"cvsStoreAddress":{"type":["string","null"],"maxLength":1000},"cvsStorePhone":{"type":["string","null"],"maxLength":1000},"storeSelectedBy":{"type":["string","null"],"enum":["admin","manual","customer_default",null]}}}
...
    CatalogCreateInput: {"type":"object","additionalProperties":false,"properties":{"name":{"type":"string","maxLength":256},"weightGrams":{"type":"string","pattern":"^\\d{1,10}(?:\\.\\d{1,2})?$"},"categoryId":{"type":["integer","null"],"minimum":1,"maximum":2147483647,"multipleOf":1},"status":{"type":"string","enum":["NORMAL","DISCONTINUED","ARCHIVED"]},"imageUrl":{"type":["string","null"],"maxLength":2048,"pattern":"^https?://[^\\s]+$"},"internalNote":{"type":["string","null"],"maxLength":2000},"preferredRouteLabel":{"type":["string","null"],"maxLength":256},"lastUsedTripRouteId":{"type":["integer","null"],"minimum":1,"maximum":2147483647,"multipleOf":1},"defaultPricingTemplateId":{"type":["integer","null"],"minimum":1,"maximum":2147483647,"multipleOf":1},"defaultShippingProfileId":{"type":["integer","null"],"minimum":1,"maximum":2147483647,"multipleOf":1},"defaultDepartmentStoreFeeRate":{"type":["string","null"],"maxLength":32,"pattern":"^\\d{1,18}(?:\\.\\d{1,12})?$"},"barcode":{"type":"string","maxLength":128},"barcodeStatus":{"type":"string","enum":["REAL","NONE"]},"forceCreate":{"type":"boolean"},"originalPriceJpy":{"type":"string","maxLength":32,"pattern":"^\\d{1,18}(?:\\.\\d{1,12})?$"},"adjustmentMode":{"type":"string","enum":["NONE","RATE","MANUAL"]},"adjustmentRate":{"type":["string","null"],"maxLength":32,"pattern":"^\\d{1,18}(?:\\.\\d{1,12})?$"},"effectiveCostJpy":{"type":["string","null"],"maxLength":32,"pattern":"^\\d{1,18}(?:\\.\\d{1,12})?$"},"adjustmentReason":{"type":["string","null"],"maxLength":2000},"observedAt":{"type":["string","null"],"pattern":"^(?!0000)(?:\\d{4}-(?:(?:01|03|05|07|08|10|12)-(?:0[1-9]|[12][0-9]|3[01])|(?:04|06|09|11)-(?:0[1-9]|[12][0-9]|30)|02-(?:0[1-9]|1[0-9]|2[0-8]))|(?:\\d{2}(?:0[48]|[2468][048]|[13579][26])|(?:[02468][048]|[13579][26])00)-02-29)$"},"supersedesCostRecordId":{"type":"integer","minimum":1,"maximum":2147483647,"multipleOf":1},"reasonCode":{"type":"string","minLength":1,"maxLength":64},"reasonText":{"type":["string","null"],"maxLength":2000}},"required":["name","weightGrams","barcode","barcodeStatus","originalPriceJpy"]}
...
    PricingV2PreviewInput: {"type":"object","additionalProperties":false,"properties":{"templateId":{"type":"integer","minimum":1,"maximum":2147483647,"multipleOf":1},"shippingProfileId":{"type":"integer","minimum":1,"maximum":2147483647,"multipleOf":1},"tripRouteId":{"type":"integer","minimum":1,"maximum":2147483647,"multipleOf":1},"originalPriceJpy":{"type":["string","null"],"maxLength":128,"pattern":"^[+-]?\\d+(?:\\.\\d+)?$"},"effectiveCostJpy":{"type":["string","null"],"maxLength":128,"pattern":"^[+-]?\\d+(?:\\.\\d+)?$"},"weightGrams":{"type":["string","null"],"maxLength":128,"pattern":"^[+-]?\\d+(?:\\.\\d+)?$"},"exchangeRate":{"type":["string","null"],"maxLength":128,"pattern":"^[+-]?\\d+(?:\\.\\d+)?$"},"generalFinalPriceTwd":{"type":["string","null"],"maxLength":128,"pattern":"^[+-]?\\d+(?:\\.\\d+)?$"},"vipFinalPriceTwd":{"type":["string","null"],"maxLength":128,"pattern":"^[+-]?\\d+(?:\\.\\d+)?$"},"costAdjustmentRate":{"type":"string","maxLength":128,"pattern":"^[+-]?\\d+(?:\\.\\d+)?$"},"departmentStoreFeeRate":{"type":"string","maxLength":128,"pattern":"^[+-]?\\d+(?:\\.\\d+)?$"},"purchasePaymentFeeRate":{"type":"string","maxLength":128,"pattern":"^[+-]?\\d+(?:\\.\\d+)?$"},"routePaymentFeeRate":{"type":"string","maxLength":128,"pattern":"^[+-]?\\d+(?:\\.\\d+)?$"},"lossProtectionTwd":{"type":"string","maxLength":128,"pattern":"^[+-]?\\d+(?:\\.\\d+)?$"},"targetMarginRate":{"type":"string","maxLength":128,"pattern":"^[+-]?\\d+(?:\\.\\d+)?$"},"templateCode":{"type":"string","enum":["GENERAL","LIVE","PERFUME","CUSTOM"]},"isTransportCostExempt":{"type":"boolean"},"thresholds":{"type":"object","additionalProperties":false,"required":["loss","low","medium"],"properties":{"loss":{"type":"string","maxLength":128,"pattern":"^[+-]?\\d+(?:\\.\\d+)?$"},"low":{"type":"string","maxLength":128,"pattern":"^[+-]?\\d+(?:\\.\\d+)?$"},"medium":{"type":"string","maxLength":128,"pattern":"^[+-]?\\d+(?:\\.\\d+)?$"}}}}}
...
    ProductSpec:
      type: object
      required: [name, values]
      properties:
        name:
          type: string
        values:
          type: array
          items:
            type: string
    Product:
      type: object
      required:
        [id, storeId, name, price, inventory, shareToken, isActive, createdAt]
      properties:
        catalogProductId: {"type":["integer","null"]}
        weightGrams: {"type":["string","null"]}
        originalPriceJpy: {"type":["string","null"]}
        effectiveCostJpy: {"type":["string","null"]}
        pricingTemplateId: {"type":["integer","null"]}
        internationalShippingProfileId: {"type":["integer","null"]}
        id:
          type: integer
        storeId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        specs:
          type: array
          items:
            $ref: "#/components/schemas/ProductSpec"
        inventory:
          type: ["integer", "null"]
        imageUrl:
          type: ["string", "null"]
        shareToken:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
        orderDeadlineAt:
          type: ["string", "null"]
          format: date-time
        internalNote:
          type: ["string", "null"]
        skuCode:
          type: ["string", "null"]
        storageTemp:
          type: ["string", "null"]
          enum: [ambient, chilled, frozen, null]
        shelfLife:
          type: ["string", "null"]
        weightKg:
          type: ["number", "null"]
        categoryId:
          type: ["integer", "null"]
        costJpy:
          type: ["number", "null"]
        isTransportCostExempt:
          type: boolean
        tripRouteId:
          type: ["integer", "null"]
    ProductInput:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: number
          minimum: 0
        specs:
          type: array
          items:
            $ref: "#/components/schemas/ProductSpec"
        inventory:
          type: integer
        imageUrl:
          type: string
        orderDeadlineAt:
          type: string
          format: date-time
        internalNote:
          type: string
        skuCode:
          type: string
        storageTemp:
          type: string
          enum: [ambient, chilled, frozen]
        shelfLife:
          type: string
        weightKg:
          type: number
          format: float
        categoryId:
          type: integer
        costJpy:
          type: number
          minimum: 0
        isTransportCostExempt:
          type: boolean
        tripRouteId:
          type: integer
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        price:
          type: number
          minimum: 0
        specs:
          type: array
          items:
            $ref: "#/components/schemas/ProductSpec"
        inventory:
          type: integer
        imageUrl:
          type: string
        isActive:
          type: boolean
        orderDeadlineAt:
          type: ["string", "null"]
          format: date-time
        internalNote:
          type: ["string", "null"]
        skuCode:
          type: ["string", "null"]
        storageTemp:
          type: ["string", "null"]
          enum: [ambient, chilled, frozen, null]
        shelfLife:
          type: ["string", "null"]
        weightKg:
          type: ["number", "null"]
        categoryId:
          type: ["integer", "null"]
        costJpy:
          type: ["number", "null"]
          minimum: 0
        isTransportCostExempt:
          type: boolean
        tripRouteId:
          type: ["integer", "null"]
...
    PaymentMethod:
      type: ["string", "null"]
      enum: [cash, bank_transfer, line_pay, other, null]
    PaymentStatus:
      type: string
      enum: [unpaid, pending, partially_paid, paid, refunded, failed]
    ShippingMethod:
      type: ["string", "null"]
      enum: [self_pickup, convenience_store, home_delivery, other, null]
    ShippingStatus:
      type: string
      enum:
        [
          not_shipped,
          preparing,
          shipped,
          arrived,
          picked_up,
          returned,
          cancelled,
        ]
    Order:
      type: object
      required:
        [
          id,
          productId,
          storeId,
          publicToken,
          buyerName,
          buyerPhone,
          pickupMethod,
          quantity,
          totalPrice,
          creditSpent,
          payableAfterCredit,
          status,
          remainingAmount,
          createdAt,
        ]
      properties:
        id:
          type: integer
        productId:
          type: ["integer", "null"]
        orderItems:
          type: array
          items:
            type: object
            additionalProperties: true
        storeId:
          type: integer
        productName:
          type: ["string", "null"]
        publicToken:
          type: string
        buyerName:
          type: string
        buyerPhone:
          type: string
        pickupMethod:
          type: string
        notes:
          type: ["string", "null"]
        specValues:
          type: object
        quantity:
          type: integer
        unitPrice:
          type: number
        shippingFee:
          type: number
        totalPrice:
          type: number
        orderTotal:
          type: number
        creditSpent:
          type: string
          pattern: "^-?\\d+(?:\\.\\d+)?$"
        payableAfterCredit:
          type: string
          pattern: "^-?\\d+(?:\\.\\d+)?$"
        remainingAmount:
          type: string
          pattern: "^-?\\d+(?:\\.\\d+)?$"
        discountAmount:
          type: integer
        discountNote:
          type: ["string", "null"]
        status:
          type: string
          enum:
            [
              pending,
              awaiting_payment,
              preparing,
              shipped,
              completed,
              cancelled,
            ]
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        paidAmount:
          type: ["number", "null"]
        paymentNote:
          type: ["string", "null"]
        shippingMethod:
          $ref: "#/components/schemas/ShippingMethod"
        shippingStatus:
          $ref: "#/components/schemas/ShippingStatus"
        recipientName:
          type: ["string", "null"]
        recipientPhone:
          type: ["string", "null"]
        recipientAddress:
          type: ["string", "null"]
        storeCode:
          type: ["string", "null"]
        storeName:
          type: ["string", "null"]
        cvsStoreAddress:
          type: ["string", "null"]
        cvsStorePhone:
          type: ["string", "null"]
        storeSelectedBy:
          type: ["string", "null"]
        storeSelectedAt:
          type: ["string", "null"]
        trackingCode:
          type: ["string", "null"]
        trackingProvider:
          type: ["string", "null"]
        shippingNote:
          type: ["string", "null"]
        internalNote:
          type: ["string", "null"]
        createdAt:
          type: string
...